+++ internal/domain/movement/api/handler.go
@@ -30,6 +30,7 @@ func NewMovementHandlers(r *gin.Engine, srv service.Movement) {
 	movementGroup.POST("/", handler.Add())
 	movementGroup.POST("/simple", handler.AddSimple())
 	movementGroup.POST("/:id/pay", handler.Pay())
+	movementGroup.POST("/:id/pay/revert", handler.RevertPay())
 	movementGroup.PUT("/:id", handler.Update())
 	movementGroup.DELETE("/:id", handler.Delete())
 	movementGroup.GET("/period", handler.FindByPeriod())
@@ -157,6 +158,37 @@ func (h handler) Pay() gin.HandlerFunc {
 
 		paid, err := h.service.Pay(c.Request.Context(), id, userID)
 		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+		if err != nil {
+			log.Printf("Error: %v", err)
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+		c.JSON(http.StatusOK, paid)
+	}
+}
+
+func (h handler) RevertPay() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		userID, err := authentication.GetUserIDFromContext(c)
+		if err != nil {
+			log.Printf("Error: %v", err)
+			c.JSON(http.StatusUnauthorized, err)
+			return
+		}
+
+		idParam := c.Param("id")
+
+		id, err := uuid.Parse(idParam)
+		if err != nil {
+			handlerError(c, model.BuildErrValidation(fmt.Sprintf("id must be valid: %s", idParam)))
+		}
+
+		paid, err := h.service.RevertPay(c.Request.Context(), id, userID)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
 			return
 		}
 		if err != nil {